requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
decorator==5.1.1
distro==1.9.0
Django==5.1.4
django-cors-headers==4.6.0
django-json-widget==2.0.1
fastapi==0.115.6
ffmpeg-python==0.2.0
frozenlist==1.5.0
future==1.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
importlib-metadata==4.13.0
jiter==0.8.2
kombu==5.4.2
language_tool_python==2.8.1
moviepy==2.1.1
multidict==6.1.0
numpy==2.2.1
openai==0.28.0
opencv-python==4.10.0.84
pillow==10.4.0
proglog==0.1.10
prompt_toolkit==3.0.48
propcache==0.2.1
psycopg2==2.9.10
pydantic==2.10.3
pydantic_core==2.27.1
pydub==0.25.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.19
pytz==2024.2
redis==5.2.1
requests==2.32.3
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
SpeechRecognition==3.12.0
sqlparse==0.5.3
starlette==0.41.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
vine==5.1.0
wcwidth==0.2.13
wheel==0.45.1
yarl==1.18.3
zipp==3.21.0
